+++ apps/@examples/modal-custom/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @examples/modal-custom
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [80cd31924]
+  - @dopt/react-rich-text@1.1.0
+  - @dopt/react-modal@1.0.7
+  - @dopt/react@3.0.4
+
 ## 0.0.2
 
 ### Patch Changes