+++ apps/@examples/modal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @examples/modal
 
+## 0.0.31
+
+### Patch Changes
+
+- @dopt/react-modal@1.0.7
+- @dopt/react@3.0.4
+
 ## 0.0.30
 
 ### Patch Changes