+++ apps/@examples/tutorial/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @examples/tutorial
 
+## 0.0.11
+
+### Patch Changes
+
+- @dopt/react@3.0.4
+
 ## 0.0.10
 
 ### Patch Changes